requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-community
langchain-google-genai
langchain-experimental
huggingface_hub==0.14.1
-e .